streamlit/requirements.txt
streamlit>=1.36
groq>=0.8.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1

